disallow edits on source=digitalglobe tag on new roads
Summary: The caveat for this implementation is that if a user manually adds a road and set source=digitalglobe, they won't be able to update it. but I think that's a really rare case, and we can add some extra checks to make sure it's ML road later.

Test Plan:
manually added source=digitalglobe on a road and verified it's not change able or deletable

for highway=living_street, clicking on the tag will open dropdown:
{F155783342}

for source=digitalglobe, clicking on the tag has no effect:
{F155783343}

Reviewers: crk, jeffunderwood

Reviewed By: crk

Subscribers: clarkben

Differential Revision: https://phabricator.intern.facebook.com/D14915244

Tasks: T42483120

Signature: 14915244:1555786041:ae829fdc1fd81757199f5f36d0d3a2323a4eaa6a