Ensure that newly added FBRoads only snap to nodes that have not been moved significantly.
Summary:
This is an attempt to prevent rubber-banding of newly added ways.

Two code paths are affected:

User adds an ML road that has a node attachment with an existing OSM Road, and:

User adds an ML road that attaches to another ML Road node (not an existing way in the graph!).

In these cases, we used to overwrite any edits made to the existing node attachments (if the user had moved them around already), which caused some weird behavior.

This will only allow such snapping if the movement of the nodes is <1meter. Otherwise, it will not attempt to snap.

One downside to this approach: If three ML roads all share a single node attachment in a Y-intersection, and the user adds one of those ways to the graph and then moves the intersection node, the next two ML roads will not attach to one another when they're added.

Test Plan:
Snapping ML roads to other ML Roads
{F157558414}

Snapping ML Roads to existing OSM Roads (with conn tag)
{F157558419}

Reviewers: crk, gaoxm

Reviewed By: crk

Differential Revision: https://phabricator.intern.facebook.com/D15208088

Tasks: T43869722

Signature: 15208088:1557432593:a02868c8230859c1bb651d5812dbad92eeb8b185