Fix RapiD toggle accuracy and restyle "off" state
Summary:
Show button state as a greyed-out "off" rather than as disabled, since it is
still interactive.

It was previously possible for the button state to get out of sync with whether
the RapiD layer was displayed. I think this was because of how the layer had a
non-canceled transition that would modify the "display" property at the end of
hideLayer(). We actually don't want the animation, and now the state updates
are immediate and the button stays in sync.

Test Plan:
{F159362362}

Reviewers: gaoxm, jeffunderwood, clarkben

Reviewed By: gaoxm

Differential Revision: https://phabricator.intern.facebook.com/D15451099

Signature: 15451099:1558539053:112b8c7ca742542c8176e5afdc54fd5558982e2e