Add new icon for cycle highway tag operation.
Summary:
@crk noticed that there is an edit hotbar when you right click on a road- and that it was showing an empty icon space where the 'cycle highway tags' operation should be!

I took the free-version fontawesome 'tags' svg and doctored it to fit the viewbox of our icons, as well as adding some 'highway divider' lines to it.

Screenshot of the icon in action:
{F158793684}

Icon itself:
{F158793708}

Test Plan: Make sure the icon appears in the menu!

Reviewers: crk, gaoxm, nickholroyd

Reviewed By: gaoxm

Subscribers: crk

Differential Revision: https://phabricator.intern.facebook.com/D15357893

Signature: 15357893:1558559246:22215212c2df7b439fcc569bb4457416039565b1