Next/previous background keyboard shortcuts v3
Binds , to previous background in list, and . to next background in
list. The thinking to my mind is that those are under the < and >
characters on a US keyboard, but that may not be international enough.

This takes the current list of backgrounds, in displayed order and
allows you to iterate through it forwards or backwards. It will only do
the custom background if there's a custom background defined.