Favorite background layers
You can now toggle stars on the far right side of the background picker
which will make the backgrounds favorited and put up higher on the list
than any of the others. Combined with the next/previous keybindings,
this makes for very fast comparisons of your favorite background layers.
The favorites will be stored in prefs and available next time you use
the system. Also use a new svg for best imagery instead of the star
(since the star now means favorite).