Next/previous background keyboard shortcuts
Brought over the v3 work to now live in 2.18.x world

Binds , to previous background in list, and . to next background in
list. The thinking to my mind is that those are under the < and >
characters on a US keyboard, but that may not be international enough.

One open question for me is the addition of the keybinding work inside
background_list.js, rather than where it seems like most of them are
occuring (in init.js). If we wanted, I could export the two functions
that need to be called and just put the keybindings in init.

This takes the current list of backgrounds, in displayed order and
allows you to iterate through it forwards or backwards. It will only do
the custom background if there's a custom background defined.