Add aws_deploy npm action to build workflow
Added the scripts/aws_deploy.py script which will run at the end of
github builds and push stuff up to s3, so that you can easily get builds
to reviewers or anyone else without issue.

Needs to have some secrets set up on github to work, documentation is at
the top of the python script.