Need to include the SDK things used by tests in the bundle
We can't actually "import thing from thing" in the test code, becasue this is
code that is supposed to run in a browser.  Normally the tests run in PhantomJS,
and this did somehow work, but I dont know how.

If you start a local server and run the tests in a modern browser at http://127.0.0.1:8080/test/
they should all work there too. This is actually a great way to debug and understand things better,
because doing it that way you get the benefit of using a real debugger.