Need to call updateText on the new Texts so they are ready for reuse
Normally updateText is called automatically when the Text actually renders as part
of the scene graph, and this is necessary to compute the size of the canvas and
prepare its texture.

If we want to reuse the text, like we do on a SimpleRope, it was getting
created with the wrong size and resolution when we skiped the setup step.