Add map.effectiveZoom() function
The "effective" zoom can be more useful for controlling the experience of the user.
This zoom is adjusted by latitude.

You can think of it as "what the zoom would be if we were editing at the equator"

For example, if we are editing in Murmansk, Russia, at about 69° North latitude,
a true zoom of 14.6 corresponds to an effective zoom of 16.

Put another way, even at z14.6 the user should be allowed to edit the map,
and it should be styled as if it were z16.