Updates to the geolocator control and PixiLayerMapUI class
The MapUI class was kind of hardcoded to only show the geolocation features,
this commit sticks the geolocation features in their own container so that
MapUI layer can be used for other things.

The reason I'm doing this now is becuase I want to stick the tile debugging
grid there. It's currently on the layer with the background imagery, so
everything gets drawn over it.