Disable texture mipmapping
Mipmapping will automatically generate downscaled copies of power-of-2 textures
at a small performance penalty.  This would be used for if we wanted to render
a downscaled copy of texture later.

But for our application, we almost always want to display the textures at
close to their native resolution, so the mipmapping is probably unnecessary,
especially as we are constantly swapping background imagery tiles in and out
of view.

This commit also sets the default resolution to 2, which generates
retina-ready textrures.  Most places that we do this, (filters, text,
rendertextures) are already set to 2, this just locks in the default in
case we forget to set it elsewhere.