Standardize on `this.context`, not `this._context`
Before we were inconsistent and using both names in different parts of the code.
(though the non-undersored one seemed more popular)

`context` is really used as an internal property that should never change and
doesn't need to be exposed to code outside class, but it's shared by everything,
so it seems fine to drop the underscore.