Comment out `searchLimited` and `partitionViewport` in photo services
The purpose of this code was to prune the results so that fewer markers would
get rendered.

The code accomplished this by splitting the viewport up into a lot of tiny square regions,
performing a rbush search for each region, and returning only a limited number of markers
in each region.

This code was pretty fast, but still did introduce a bit of performance penalty to do all
that searching and filtering.  But my bigger issue is that it's not really the service's
job to do this.  The renderer itself should have access to the complete result set and
then decide what to do with the markers.