Use a `v` version on objects we want to watch for changes
I considered Proxy/Reflect, Object.defineProperty, and a few other more serious
ways of adding props to things but this is simple and we already have a `.v`
property in a few other places.

I am a tiny bit concerned about the performance hit from changing an Object's "shape"
after it has been created, but if this becomes a real issue we can someday create like
actual VersionedSet and VersionedMap or something.