Handle the possiblilty that a feature can change its type (closes #560)
In OSM it's possible for a feature to convert from line->polygon or polygon->line
if its tags change.  Here we handle that possibility where the feature was already
created and exists in the scene as something else, and it needs to be recreated.

I added code to the point, vertex, midpoint functions too, just in case,
but I dont think this can happen.  OSM Nodes always stay PixiPoint features.