Move `parent` constructor argument to class property continuation of 979fbd012
Going forward I'm going to try to keep constructor arguments as things that
never change during the lifetime of the object, and class properties as things
that can change anytime (and where we handle those changes thoughtfully)
Long argument lists are confusing to work with.

This change also makes it a little bit easier to re-parent a feature.
We currently do this when the layer stacking changes (e.g. bridge-tunnel).