Put strokes back into PixiFeaturePolygon
This means polygons are drawing their strokes again so things look nice.
This probably also means that for OSM multipolygons, the strokes are being
drawn twice, once by PixiFeatureLine and once by PixiFeaturePolygon.
I think the lines are being drawn on top and seem to be hover/selectable.
This is probably good enough for now.