Make the final step in each chapter return a resolved Promise
This one has me a bit stumped.  Previously, the runAsync function would run
the final step, but then enter the .catch() block and retry forever on the
penultimate step.

Running the final step in the .then() and having no return value,
should be the same as if .then() return a Promise resolved to `undefined`.
And stepping into runAsync with no argument should hit the "guess we're done"
condition and resolve.  I'm not sure why we'd catch anything.

But anyway, it's no problem to just return a Promise.resolve() so that's what
we'll do.