When retrying the "road does not intersect" step, just use an Ok button
The old way would delay for a bit and then push the user back to the previous step.
A while ago we indended to replace these "forced pushes" with Ok buttons because we
felt the user should move at their own speed, and it makes the code a bit simpler.