If chaining a promise after a delay, check preconditions in the promise
This was causing issues where the user could switch modes while the delay was
happening. An easy way to trigger this was to hit escape twice quickly during a
step that required a thing to be selected (like a "search for this preset" step).

This would essentially bail out of select mode, restarting the step, which
would try to keep the thing selected, then the second escape would bail
out of it again, but after the short delay the reveal would not be able to
reveal the preset search input.