When resetting history to a checkpoint, dispatch 'change' with difference
This is necessary so that if the tutorial resets the history back to a
previous state, we notifiy the renderer what nodes need a redraw.

Was mostly noticable before in the steps such as if the user didn't square
the building properly, we'd ask them to redo it. Without this, it would reset
the building to it's previous shape, but not actually redraw it.