Operations cleanup and ES6ify, initial pass
This commit is a rough pass over all the operations to simplify and use more modern code.

It also changes all the operations so that they skip some disabled reasons for
geometry that was newly drawn by the user.  For example, users should not see a
message about something "not enough is currently visible" if they just drew it and
it hasn't been uploaded to OSM yet.

(closes #822)