Don't set `interactive` on everything that's !activeData (closes #812)
This fixes the "Nodes should not be clickable/selectable in the RapiD Roads layer"
issue, and possibly some other things.  I'm leaving the code commented out for now
because I need to remember what it was for originally, and the new Pixi 7.2 release
has some more granular interactivity settings that I need to dig into.