When snapping to a way, don't snap to an active segment It's not perfect, but it's ok enough for now
This adds a workaround for something that used to work a while ago when we had
css/svg renderer and specific touch targets.  When drawing or dragging nodes,
you can allow a way to target itself, but not adjacent to the segment that is
being drawn/dragged - it needs to be further along the way to snap.

This doesn't work perfectly.  A better solution would be to draw each segment
of way as its own geometry that can be targeted with Pixi's event system.