Make "zoom to selection" work for any selection, not just OSM (closes #866)
Now if a mode exposes an `extent` property, the "Zoom to selection" tool will
zoom in on it if the user clicks the button or presses 'Z'.  Clicking or
pressing 'Z' again will pop back out to whatever the user was looking at before.

This works for both the legacy OSM select mode and the new select mode that
handles notes, custom data, QA feature, and Rapid data.