fileFetcher -> DataLoaderSystem
Rapid does work, but now the tests are a mess.

This is because the interactions between the localizer and everything else get
complicated, as creating a context will start loading data and the fileFetcher
isn't global anymore.  The tests did have a lot of code in spec_helpers to
prevent files from being fetched, that doesn't work now.

This can start getting easier after we change the localizer similarly.