Fix the init ordering so that MapSystem gets first hashchange
Many things in Rapid depend on the `map=zoom/lat/lon` being set correctly,
so it's important that MapSystem gets the first hashchange event to set it up.

This means that code that listens for UrlHashSystem.hashchange event,
needs to add their listener after MapSystem.initAsync has finished.

Note that EventEmitter doesn't support prependEventListener like Node does.
If it did, that would be another way to ensure the correct ordering.