Modernize uiEntityEditor, fix issue with _onChangeListener (fixes #1084, related to the change in: #989)
This was a very weird issue!

When reset happens we throw away parts of the UI and let d3 recreate them.
This means that we'd see weird behavior _after_ the user saved their changes.

The effect was that the uiEntityEditor component was no longer listening to history change events:
`  context.systems.edits.on('change', _onChange);`

but had also already set the _init flag to true.

This meant that the uiEntityEditor component - and all its child sections and fields -
were always working off of whatever the tags looked like when the component first
rendered, and it would never receive updates or render properly as the user typed into
fields, or clicked the various buttons to manipulate the tags.