Continuing core refactor, particularly for EditSystem (followup from ca9be6003, also see #961)
Continuing to remove editing and localization operations from Context.
The idea here is to remove "magic" from the code.
These functions were just redirects from the Context to the EditSystem,
so using them obscured what was really going on.  It was easy to get
confused about "which graph" the code was really using.

- Removed `context.graph`, `.entity`, `.hasEntity`
  Code must go through the EditSystem to do these things now.

- Convert several EditSystem accessor methods to ES6 getters
  - `.graph()`  - removed
  - `.base`     - now a property
  - `.current`  - added
  - `.current` and `.base` now return the `Edit`, which has `.graph` as a property.
    This makes it easier to compare the current and base graph, and just more consistent flow.

- Also WIP on replacing `context.t` and `.tHtml`
  Code to do these things should go through the LocalizationSystem
  There is quite a lot of this, and it will take a while.