Move responsibility for saving backups into the EditSystem
Previously we had methods in context that would perform edits "withDebouncedSave".
This should just be the EditSystem's responsibility.

This commit renames a bunch of things for clarity (we call them backups now)
and makes the EditSystem responsibile for its own backing up and restoring,
as well as registering `beforeunload` handler and handling the mutex lock.