Replace undo/redo with historychange in Move and Rotate modes
Added `.hasWorkInProgress` so other code can decide whether to finalize or not.
(This is because we can't easily compare graphs - I initially considered code
 like `editor.current.graph === editor.stable.graph` but because current is
 always a copy it will never be strict equals.)

Adjusted how we reassign _lastStableGraph and _lastCurrentGraph in
_emitChangeEvents, to avoid situation where an event receiver does something
like editor.rollback() and kicks off an infinite loop / event storm.

Changed undo so that it will rollback work-in-progress first before backing out
the history.  This is intuitively how users would expect undo to work.