Start working towards making mode options serializable
At some point we will want to start recording in the edit history:
- the mode that the user is in, and
- whatever options (selected IDs) they used in that mode

This is to support situations where we are selecting a mix of OSM and non-OSM
features and then doing edits to them.  Currently we are assuming everything
in the edit history (selectedIDs) is an OSM feature.

If we want to support where users are merging data from multiple sources and
then be able to undo back out of those edits, it would be helpful for everything
we need to be stored in the history, and backed up like other edits.

We're not there yet, but this commit takes us closer to being able to do that.