Allow commit to accept options include annotation and selectedIDs
Previously we were just taking the `context.selectedIDs()` - this change gives the
calling code more control over what ids will actually be stored with the history entry.

It's useful for situations like when you are adding a point - the point will not be
selected until _after_ the commit has happened, but we'd still want to undo back into
the state where that point exists and is selected.

This change also includes a separate event for `statechange` which fires a bit
differently than `historychange`.  It's for situations like undo/redo/restore where
we want to adjust the map transform and the selection.

We previously had this code in a few places - now it just lives in MapSystem's
`statechange` handler.  I'm not sure whether it really belongs there.  This may
end up being more EditSystem's responsibility eventually.