Renames `current` -> `staging`, `rollback` -> `revert` (see #1156)
Also the events are now `stagingchanged` and `stablechanged`

Also fixed a few issues with the Validator `validateAsync` that the tests exposed.
Basically, it was returning a resolved Promise even though the validation had
started and was in progress.

Also fixed failing test in Difference.test.js constructor test.
It was trying to diff 2 Graphs that had different bases - this is not a thing we do in
the appliction, so there's no expectation that Difference should consider base entities.
(and doing so would be a pretty big performance penalty - the base entities get huge!)