Fix issue with map failing to fit to entity in `id=` param (closes #1108, closes #1173)
This issue happened because, although the MapSystem contains the code to
fit the map to the selected entity - it was choosing not to do it.

The cause was the this line of code:
  `entityExtent.percentContainedIn(this.extent()) < 0.8`

This evaluates `true` when if the map happened to be zoomed out,
it clearly contains the whole entity, so the fitEntities line was skipped.

I added a check to see whether the entity would be considered "too small"
and renamed several of the variables to make the code intent clearer.