Add ability for the renderer to reset
This adds a `reset()` method to the scene/layers to clear out all layer state on a
reset.  It's not good to hang on to this stuff across resets because the entities
data or ids will be different.

This also updates `MapSystem.resetAsync()` to be responsible for calling
`reset()` on the scene and then schedule an immediate render on the next tick.