Convert EditSystem `fromJSON` -> `fromJSONAsync` (closes #1074)
Hopefully fixes this longstanding issue!

The main issue with this `fromJSON` function is a common source of bugs in our code:
 treating async things as if they are sync, or vice versa.

`fromJSON` was especially gnarly beacuse it can work either way - normally it would
be sync, unless it needs to go looking for missing OSM entities, then it becomes async.

Now that more parts of the code are async/promisified - this code should be too.
It's now a Promise-returning function (hence the "async" in the name) intended to be
chained after a `context.resetAsync()` to ensure that we are starting with a clean slate.

This commit updates both the restore code in EditSystem, and also the code to
enter and leave the walkthrough, these are the only 2 places where we use fromJSON.