Make sure sammple Rapid data shows up in the walkthrough (closes #1227)
This was caused by recent changes to reset and editsystem.
Now the code does a `context.resetAsync` when entering and leaving the walkthrough.
When that happens, the service resets all of its state.

So the call to load the sample data (`mapwithai.merge`) needs to be chained after the `resetAsync`.
We were already doing this for the other sample OSM data, just missed doing it for the Rapid data too.

This commit also includes a few more checks within the rapid chapter to really
make sure the data is turned on when it needs to be on.
(It is possible in the walkthrough for the user to open the Rapid dialog box and uncheck it)