Fix bug causing by stale preset lists used by inputevent
The bug was causing the list to sometimes display the wrong geometries,
e.g. if you select a line, and then select a vertex, then use the preset
picker, you might still be offered the line presets when typing in the search.

This is a common issue in JavaScript when an event handler refers to a
closure variable, it will be the variable at the time the function is created.

We can fix this by being more diligent about variables that we expect to
act stateful (change later) by accounting for all the ways we expect them to
change, prefix them with `_`, reset them when needed.

In this commponent the stateful variables should always reset on any change to
entityIDs.