Having sync_crossing_tags run on any tag change is too heavy-handed (re: #1260)
This makes it difficult or impossible to manually correct tags in the raw tag
editor. We can keep the code in actionChangePreset, but remove the code from
actionChangeTags and actionUpgradeTags.

I've added more logic to actionSyncCrossingTags, so it can be used by the
validator to raise issues.

I think the ideal user experience is:
- when changing presets, apply the updates automatically
- when editing raw tags or field values, raise validation issues for mismatched values