SyncCrossingTags now includes an initial clean step (re: #1210)
This sets default values where missing and removes obvious mismatches.
This provides a lot of value, but its also the reason we had to remove the
automatic-ness of this action.  Because if the user tries changing the raw
tags, this action would overwrite their changes and user would be stuck.