Allow the outdated_tags validator to upgrade standalone crossing nodes
This validator can perform preset upgrades on standalone crossing nodes
  that are NOT attached to a parent crossing way.

If the crossing is: 1. a way or 2. a node attached to a parent crossing way, bail out.
The `ambiguous_crossing_tags` validator will take care of them.

(i.e. that parent way is the thing that will get validated thoroughly)