uiCmd shouldn't confuse up the shortcut with the display of the shortcut (re: #1272, re: #1261, re: #1255)
Windows/linux should not replace '⌘Z' with 'Ctrl+Z', it should just replace it
with '^Z' and then the `uiCmd.display` should handle how to present that to the user.

This change simplifies the `uiCmd` code, and avoids the 'Ctrl+` stuff.

It also changes `uiTooltip` to work with an array of keys, using
d3-selection.each and `uiCmd.display` to render each one.
Previous code was attempting to split the keycombo into chars, causing #1272

I still need to make sure that `tooltip.keys()` is being called consistently,
with an Array of keys - we do this sometimes but not everywhere.

Also, in 01a16a990, the '+' to zoom in broke, because it is a shifted key.
This commit fixes that issue also.