Be more careful about syncing incomplete crossing tags
Before, the code would attempt to sync tags even if we saw a way tagged something like `crossing=foo`
This is because the code assumed that it might be a railway, which can have crossings along it.

Now we check a few more tags in `_isCrossableWay`, and make a better choice
whether to either remove the tags or bail out and leave them alone.