Avoid raising `outdated_tags` warning for a crossing on unloaded tile
The `outdated_tags` validator can upgrade crossing child nodes.
The `ambuguous_crossing_tags` validator can upgrade crossing ways + child nodes.

This created a situation where we'd sometimmes display an issue to "upgrade the tags"
from `outdated_tags` but then it would quickly flash away and be replaced with a different
issue about upgrading the tags / resolving the conflict from `ambiguous_crossing_tags`

If the crossing node is on an unloaded map tile, we won't know for sure whether it
is a just a isolated crossing node or has a parent crossing way, so we should just
defer running the validation until later when that part of the map has loaded.

This is a frequent issue with other validation code too, so I've added
some more code comments to the other places where we have this sort of check.