Fix: Adding a detected feature can cause nearby features to vanish (closes #1288)
This was an artifact caused by the way-splitting behavior of our mapwithai service.

In the past we tried correcting for this by storing both `origid` and `entityid`
and considering either of those as sufficient for marking whether a way was
accepted or ignored.  Because multiple adjacent ways can share and `origid`
this made them disappear.

This commit makes a bunch of changes in how we deal with this way-splitting.
- `origid` is now an internal detail to the MapWithAIService code, it won't
  be used anywhere else in Rapid
- ways that were split by the service will be joined back together by the
  new `_connectSplitWays` function .. These ways will use the origid as
  their entityID for stability..  From outside the MapWithAIService code,
  the ways just appear as normal Rapid ways with stable ids.