Refresh edit_menu when loading new entities from the OSM API (re: #1311)
This commmit includes a change to edit menu to make it reactive
to changes in data (the operations available in the current mode).

So if called multiple times, now it will actually update instead of
just making another menu.  Also includes a `redrawEditMenu` function
in the UiSystem, for situations where we want to refresh the menu
with updated content.