Remove the `overzoom` prop from ImagerySource (closes #1321)
The original intent for this was to mark imagery sources that shouldn't be
overzoomed, because they would get blurry. However it was only set `false`
on one source, the mapbox locator overlay, and defaulted to `true` otherwise.

It wasn't intended to extend `validZoom` beyond the max zoom to do tile
requests out of range.  So because it was producing this unexpected behavior
and we weren't really using it anyway, it's ok to remove it.  Imagery will
just be blurry if the user zooms in a lot, and this is ok.