Improve `inferCrossingType` when `crossing:markings` mismatch `crossing` (closes #1298)
Previously we were only looking for "interesting" values of `crossing:markings`, then falling
back to whatever the `crossing` value contained. This means that in a mismatch situation,
it might end up choosing the wrong tag to apply when the user clicks the fix button.

I changed this code so that it now properly considers the `crossing:markings` 'yes' and 'no'
values too, before falling back to looking at the `crossing` tag.