Never show overlay when zoomed in too far (re: #1321)
This relates to the `overzoom` param that we aren't using anymore.
The overlay was the only source where it made any difference.

What would happen is that, at normal editing zooms, the user could see
a brief flash of blurry zoomed in overlay tiles, which get immediately
removed as actual OSM data streaming in.

This commit also replaces a few uses of the sourceID 'mapbox_locator_overlay'
with the convenience method `isLocatorOverlay()`.  There isn't much difference
in this approach, but it would let us more easily swap out the overlay with
a different layer someday if we need to.