This removes the extra 'reset' from ValidationSystem
This code used to listen for the reset event from `history` a long time ago,
but now that these things are Systems, they receive their own `resetAsync` calls
and this becomes unnecessary.

The one thing I'm not sure of is:  the two resets used to differ in only one
way and that's whether it reset the user's cache of ignored issues.  Because
these issues reference issueIDs, which are constructed from entityIDs, I feel
like we can't persist them across resets - any kind of reset of the graph
should clear the ignored issues too.

We'll have to think about whether there is a better way to let the user keep
their ignored issues or manage this list somehow.