WIP on rotation, attempting to provide "screen" and "surface" coords`
The hope here was that screen coords are relative to the viewport
and surface coords include rotation.

I'm having trouble though, because hit testing is done by Pixi and it
uses the "global" coords and worldTransform, which are not going to have
rotation included.  This means I need to change my approach and include
rotation in the stage, not the supersurface.