Simplify the code around working with viewports
There's a lot going on in here.

Before, we had a class called Projection, it lived in Context
It handled projections, but was starting to include more things like map dimensions.
https://github.com/rapideditor/rapid-sdk/issues/266

Now, the Viewport class does that, and it still lives in Context.
We still had a lot of code that was calling into MapSystem to get things that is
really the viewport's job.  This commit removes most of that indirection.

e.g.
map.centerLoc -> viewport.centerLoc
map.center -> viewport.centerLoc
map.extent -> viewport.visibleExtent
map.dimensions -> viewport.visibleDimensions
map.zoom -> viewport.zoom
map.transform -> viewport.transform

Think of the Viewport as - the rectangle of Earth that the user is looking at
Think of the MapSystem as - everything needed to interact with the slippy map

There are MANY parts of the code that don't need a full slippy map, but still
should be able to know what's going on in the Viewport.  Things like service
code to fetch data in view, validator code to get issues in view, imagery code
to know what sources are valid where the user is looking.

As an additional benefit, testing the code has gotten simpler, because many
tests that previously needed to mock the MapSystem don't need to do that
anymore.