Restore overlay / editMenu
This also adds an experiment to the UiSystem.resize function:
Previously, the map surfaces were anchored to the top left of the main-map.
Now, the map surfaces are centered in a CSS Grid, to support rotation around the center.

Because the surfaces are now guaranted to be centered within their container, we can just
extend their dimensions a little bit, so that as the user pans the map, they are less likely
to see seams at the edges of the map.  (We're drawing a few extra pixels of "overscan")