Don't use 'inline' for the interaction icons - it made them too small
Our icon css rules are very confusing, and we should clean them up eventually.
- 'inline' makes the icon very small, but makes it look like it fits exactly
  within the text it appears next to.  This is good for the outbound link icons,
  or things like "< Previous", "Next >"
- 'pre-text' just adds a small margin next to the icon so it doesn't bump up
  against whatever text is it next to.

In some cases we need 'inline' or 'pre-text' or neither of these.  There's
been a lot of cargo culting the icon styles, so there are many places in
the code that uses these rules when they don't really need it.