Don't `skipUpdates` when calling `origin.toGlobal()` to get labelOffset
I don't completely understand this, but..
We were seeing label rendering artifacts when using `skipUpdates = true`.
What seemed to happen was that it used a stale transform for computing the `labelOffset`
So the labels would - sometimes at higher zooms and when panning - be shifted away from
where they were supposed to be, but the pop back into place later.