Avoid tiny rotations from creeping into the transform
It was possible before to end up with a rotation like 0.0001, it would not appear
in the url hash, and the reset bearing button would not be able to remove it.
This commit includes some defensive code to try to avoid these situations.

Also modifies the code in a few places to try to follow the Pixi convention of
using "angle" for values in degrees, and "rotation" for values in radians.