When deciding which tiles to fetch, round up for sharper imagery (closes #485)
This solves the issue of the minimap labels looking bad and overzoomed
but also would be useful for the main map imagery tiles too.

Rounding down will always result in overzoomed tiles, and this was noticeably blurry
especially around the .5 rounding cutoff, so probably better to just round up and
display the sharpest imagery possible.  The downside is that we'll fetch more tiles
than before when fraction zooms is between .0 and .5 .