Fix UICurtain to calculate coordinates based on supersurface (re: #1363)
Since the map rotation changes, the supersurface may be expanded or rotated
beyond the map surface, so we need to account for that when determining the
size of any Extent on the map that the walkthrough is revealing.

This is not perfect, as the reveal can only do screen-aligned rectangles, so
if you rotate the map, it will shrink the reveal rectangle into a tiny sliver.

This is kind of ok though because none of the steps in the walkthrough where
we reveal a map Extent, we are expecting the user to rotate the map, and if
the user does rotate the map out of curiosity, they can just rotate it back
if they find the revealed rectangle unusable.