ImagerySystem can now deal with embedded date in sourceID for wayback
This means that the urlhash can handle `sourceID` that look like
`background=EsriWayback` or `background=EsriWayback_2024-01-01`
This is important for preserving the Wayback imagery across refreshes.

The ImagerySourceEsriWayback now has a date getter/setter, and correctly
returns a reasonable `.imageryUsed` that includes the date if there is one.

It is not yet fetching or checking all the available dates - I need to restore that code next.