Make Esri Wayback a copy of Esri World Imagery + other fixes
Copying the Esri World Imagery props means we have less values we need to
hardcode.

This also adds a `.key` getter for sources.  It usually returns the `.id` but
for Wayback, it will include the date as well.  This is for places in the code
that need to uniquely identify different wayback sources, such as:
- ImagerySystem where it writes this value to the urlhash
- PixiLayerBackgroundTiles tile rendering code
- UIPanelBackground info panel