Force a render when toggling open a pane (background, map date, etc) (closes #1405)
The main fix here is in pane.js, to force a re-render just as the pane is
opening.  These panes already do listen for events that cause them to render,
but sometimes those renders are skipped if the content is collapsed or
invisible.
In the case of the photos overlay section, they were not receiving an event to let
them to know to redraw the checkboxes correctly.

This commit also includes a bunch of updates to source documentation
and some small style adjustments to the maproulette challenge id input field,
unrelated to the actual fix.