Improve new user user experience: What's New and Splash Screen (closes #1325)
These screens now do a better job welcoming the user to Rapid and encouraging
them to take the tutorial if they are brand new.

I commented out the First Edit screen for now.. It only nags the user to log
into OSM, which they will need to do anyway when they save.  We can improve
this one later if we want.