Update whats new screen for v2.3
I'm not totally sure whether I did this right, but we'll see whether it works
in canary.  I tried adding a HTML <video> tag that should show a .mp4 of the
map rotating, with fallback to a static .jpg image.

When I test it locally, I don't see the video play, but that might just be
because `static-server` isn't serving partial content, and it might just work
when being served from a real web server?